requirements.txt
# Production Dependencies for Buy Printz Platform
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
stripe>=7.6.0
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.1.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
supabase>=2.3.0
bcrypt>=4.1.2
openai>=1.3.0
httpx>=0.25.2
qrcode[pil]>=7.4.2

# Web Scraping Dependencies for Shipping Integration
beautifulsoup4>=4.12.2
selenium>=4.15.0
redis>=5.0.1

# Testing Dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# FastAPI includes Swagger UI built-in, no separate package needed
